Add PermissionManager and write contacts implementation
