+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="eu.ezytarget.micopi">
 
     <uses-permission android:name="android.permission.READ_CONTACTS" />
+    <uses-permission android:name="android.permission.WRITE_CONTACTS" />
 
     <application
         android:allowBackup="true"
