+++ app/src/main/java/eu/ezytarget/micopi/common/permissions/PermissionManagerCallback.kt
@@ -1,3 +1,3 @@
-package eu.ezytarget.micopi.common
+package eu.ezytarget.micopi.common.permissions
 
 typealias PermissionManagerCallback = (Boolean) -> Unit
