+++ app/src/main/java/eu/ezytarget/micopi/main_menu/MainMenuViewModel.kt
@@ -5,11 +5,12 @@ import android.content.Context
 import android.content.Intent
 import android.view.View
 import androidx.lifecycle.ViewModel
+import eu.ezytarget.micopi.common.permissions.PermissionManager
 
 class MainMenuViewModel: ViewModel() {
 
     var selectionListener: MainMenuSelectionListener? = null
-    var contactPermissionManager: ReadContactsPermissionManager = ReadContactsPermissionManager()
+    var contactPermissionManager: PermissionManager = ReadContactsPermissionManager()
     var contactPickerResultConverter: ContactPickerResultConverter = ContactPickerResultConverter()
     private var allowMultipleSelection = true
 
@@ -43,8 +44,8 @@ class MainMenuViewModel: ViewModel() {
     }
 
     private fun validatePermissionsAndSelectContactPicker(activity: Activity) {
-        if (!contactPermissionManager.hasReadContactsPermission(activity)) {
-            contactPermissionManager.requestWriteContactsPermission(activity) {
+        if (!contactPermissionManager.hasPermission(activity)) {
+            contactPermissionManager.requestPermission(activity) {
                 val permissionGranted = it
                 if (permissionGranted) {
                     selectContactPicker()
